Factor out test helpers
