+++ shared/src/types/ethereum_events.rs
@@ -358,3 +358,56 @@ pub struct EthMsgDiff {
     /// addresses of the validators who have just seen this event
     pub seen_by: Vec<Address>,
 }
+
+#[allow(missing_docs)]
+/// Test helpers
+#[cfg(any(test, feature = "testing"))]
+pub mod testing {
+    use rand::prelude::ThreadRng;
+
+    use super::vote_extensions::*;
+    use super::*;
+    use crate::types::key::{common, ed25519, SigScheme};
+    use crate::types::storage::BlockHeight;
+    use crate::types::token::Amount;
+
+    const DAI_ERC20_ETH_ADDRESS: &str =
+        "0x6B175474E89094C44Da98b954EedeAC495271d0F";
+
+    pub fn arbitrary_eth_address() -> EthAddress {
+        let bytes: [u8; 20] =
+            hex::decode(DAI_ERC20_ETH_ADDRESS[2..].as_bytes())
+                .unwrap()
+                .try_into()
+                .unwrap();
+
+        EthAddress(bytes)
+    }
+
+    pub fn arbitrary_fractional_voting_power() -> FractionalVotingPower {
+        FractionalVotingPower::new(1, 3).unwrap()
+    }
+
+    pub fn arbitrary_nonce() -> Uint {
+        123.into()
+    }
+
+    pub fn arbitrary_amount() -> Amount {
+        Amount::from(1_000)
+    }
+
+    pub fn arbitrary_block_height() -> BlockHeight {
+        BlockHeight(100)
+    }
+
+    /// This will actually generate a new random secret key each time it's
+    /// called
+    pub fn arbitrary_secret_key() -> common::SecretKey {
+        let mut rng: ThreadRng = rand::thread_rng();
+        let sk: common::SecretKey = {
+            use crate::types::key::SecretKey;
+            ed25519::SigScheme::generate(&mut rng).try_to_sk().unwrap()
+        };
+        sk
+    }
+}
